Initial pass at making a package
